requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
langchain>=0.2.14
langchain-openai>=0.1.20
streamlit>=1.36.0